refactor: facade pattern 추가해서 layer 분리
